[arp]: Refactor utility methods and fixtures
